% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{check_t_lambda}
\alias{check_t_lambda}
\title{Check t and lambda}
\usage{
check_t_lambda(t, lambda)
}
\arguments{
\item{t}{time interval starting points}

\item{lambda}{hazards}
}
\value{
TRUE invisibly
}
\description{
Check t and lambda
}
\details{
check if t and lambda have valid and compatible size, sign and order
}
