% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.miniPCH}
\alias{plot.miniPCH}
\title{miniPCH class}
\usage{
\method{plot}{miniPCH}(
  obj,
  what = c("d", "s", "h"),
  from,
  to,
  mfrow = c(1, length(what)),
  n = 1001,
  ...
)
}
\arguments{
\item{obj}{miniPCH object}

\item{what}{what to plot ("d", "p", "q", "h", "ch", "s")}

\item{from}{lower x-Axis limit}

\item{to}{upper x-Axis limit}

\item{mfrow}{plot layout defaults to all plots in one row}

\item{n}{number of points for interpolation}

\item{...}{passed on to base::plot}
}
\value{
NULL, invisibly
}
\description{
miniPCH class
}
\details{
The layout uses the mfrow argument to par and defaults to all plots in one
row. The layout can be overwritten by passing the mfrow argument, that is
passed as is to an internal call to par.
}
\examples{
my_pch <- pch_functions(c(0, 3), c(2, 0.1))
plot(my_pch)

Tint <- c(0,3)
Q <- array(
  c(
    -0.3,  0  , 0,
    0.2, -0.4, 0,
    0.1,  0.4, 0,

    -0.3,  0  , 0,
    0.2, -0.2, 0,
    0.1,  0.2, 0
  ), dim=c(3,3,2)
)

pi <- c(1,0,0)
abs <- c(0,0,1)

my_obj <- multistate_functions(Tint, Q, pi, abs)
plot(my_obj)
}
